# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MentorAlm API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get student profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update student profile (any subset of fields)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /profile/photo:
    post:
      operationId: uploadProfilePhoto
      tags: [profile]
      summary: Upload profile photo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUploadResponse"
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat message history
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
  /chat/stream:
    post:
      operationId: streamChatMessage
      tags: [chat]
      summary: Send a message and receive SSE streamed AI response
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response
          content:
            text/event-stream: {}
        "429":
          description: Daily rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/new:
    post:
      operationId: newChat
      tags: [chat]
      summary: Start a new chat session (clears history)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: New session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /roadmap:
    get:
      operationId: getRoadmap
      tags: [roadmap]
      summary: Get career roadmap
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Career roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roadmap"
        "404":
          description: No roadmap generated yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /roadmap/generate:
    post:
      operationId: generateRoadmap
      tags: [roadmap]
      summary: Generate or regenerate career roadmap using AI
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Generated roadmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roadmap"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/contacts:
    get:
      operationId: listAdminContacts
      tags: [admin]
      summary: List contact form submissions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dateOfBirth:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        educationLevel:
          type: string
          nullable: true
        board:
          type: string
          nullable: true
        schoolCollege:
          type: string
          nullable: true
        gradePercentage:
          type: string
          nullable: true
        stream:
          type: string
          nullable: true
        interests:
          type: string
          nullable: true
        strengths:
          type: string
          nullable: true
        hobbies:
          type: string
          nullable: true
        energizedBy:
          type: string
          nullable: true
        dreamCareer:
          type: string
          nullable: true
        targetColleges:
          type: string
          nullable: true
        openToAbroad:
          type: string
          nullable: true
        biggestConcern:
          type: string
          nullable: true
        familyIncome:
          type: string
          nullable: true
        parentsEducation:
          type: string
          nullable: true
        familyPressure:
          type: string
          nullable: true
        educationBudget:
          type: string
          nullable: true
        fiveYearGoal:
          type: string
          nullable: true
        alreadyTried:
          type: string
          nullable: true
        stoppingYou:
          type: string
          nullable: true
        heardFrom:
          type: string
          nullable: true
        completionPercent:
          type: number
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - completionPercent
        - updatedAt
        - createdAt
    UpdateProfileInput:
      type: object
      properties:
        dateOfBirth:
          type: string
        gender:
          type: string
        city:
          type: string
        state:
          type: string
        educationLevel:
          type: string
        board:
          type: string
        schoolCollege:
          type: string
        gradePercentage:
          type: string
        stream:
          type: string
        interests:
          type: string
        strengths:
          type: string
        hobbies:
          type: string
        energizedBy:
          type: string
        dreamCareer:
          type: string
        targetColleges:
          type: string
        openToAbroad:
          type: string
        biggestConcern:
          type: string
        familyIncome:
          type: string
        parentsEducation:
          type: string
        familyPressure:
          type: string
        educationBudget:
          type: string
        fiveYearGoal:
          type: string
        alreadyTried:
          type: string
        stoppingYou:
          type: string
        heardFrom:
          type: string
    PhotoUploadResponse:
      type: object
      properties:
        photoUrl:
          type: string
      required:
        - photoUrl
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - role
        - content
        - createdAt
    ChatMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    Roadmap:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        generatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - content
        - generatedAt
    ContactInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        message:
          type: string
      required:
        - name
        - email
        - message
    ContactMessage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - message
        - createdAt
...